Removed some warnings when building, and added another project reference to be sure it doesnt start breaking all of a sudden.

+++ ParsersTests/AOFullCharacterMessageTests.cpp
@@ -26,7 +26,7 @@ struct AOFullCharacterMessageTestFixture : public testing::Test
 
         if (ifs.is_open())
         {
-            int size = ifs.tellg();
+            int size = (int)ifs.tellg();
             ifs.seekg(0);
 
             char* buffer = new char[size];
